+++ tests/test_all_tasks.py
@@ -12,13 +12,19 @@ task_names.insert(0, "p3::wiki_qa_Is_This_True_")
 
 
 @pytest.mark.parametrize("task_name", task_names)
-@pytest.mark.parametrize("split", ["train"])
-def test_task(task_name: str, split: str):
+def test_task(task_name: str):
     task = catwalk.tasks.TASKS[task_name]
-    try:
-        instances = task.get_split(split)
-    except KeyError:
-        return
+    
+    # get the first available split for this task
+    instances = None
+    for split in ["train", "validation", "test"]:
+        if task.has_split(split):
+            instances = task.get_instances(split)
+            break
+        
+    if not instances:
+        raise KeyError("No split found for task {}".format(task_name) + " in splits {}".format(splits))
+
     for conversion in task.instance_conversions.values():
         signature = inspect.signature(conversion)
         for instance in instances[:10]: